Filter out hidden entities from players
