+++ src/components/sidebar-entity/entity-edit-row/index.js
@@ -5,8 +5,10 @@ import {
   undoDeleteChildInEdit,
   updateChildInEdit,
 } from 'store/actions/sidebar-edit.actions';
-import { isCurrentOrAncestorHidden } from 'store/selectors/entity.selectors';
-import { getEmptyHexKeys } from 'store/selectors/sector.selectors';
+import {
+  isCurrentOrAncestorHidden,
+  getEmptyHexKeys,
+} from 'store/selectors/entity.selectors';
 import EntityEditRow from './entity-edit-row';
 
 const mapStateToProps = state => ({
