+++ src/store/selectors/entity.selectors.js
@@ -1,5 +1,12 @@
 import { createSelector } from 'reselect';
-import { filter, pickBy, mapValues, zipObject } from 'lodash';
+import {
+  filter,
+  pickBy,
+  mapValues,
+  zipObject,
+  difference,
+  values,
+} from 'lodash';
 
 import {
   currentSectorSelector,
@@ -8,8 +15,11 @@ import {
   entitySelector,
   isSidebarEditActiveSelector,
   sidebarEditEntitySelector,
+  sidebarEditChildrenSelector,
 } from 'store/selectors/base.selectors';
+import { isViewingSharedSector } from 'store/selectors/sector.selectors';
 import Entities from 'constants/entities';
+import { allSectorKeys, coordinateKey } from 'utils/common';
 
 export const getCurrentTopLevelEntities = createSelector(
   [currentSectorSelector, entitySelector],
@@ -88,16 +98,36 @@ export const getCurrentEntityId = createSelector(
 );
 
 export const getCurrentEntityChildren = createSelector(
-  [getCurrentEntityId, getCurrentEntityType, entitySelector],
-  (entityId, currentEntityType, entities) => {
+  [
+    getCurrentEntityId,
+    getCurrentEntityType,
+    entitySelector,
+    isViewingSharedSector,
+  ],
+  (entityId, currentEntityType, entities, isShared) => {
     const entityChildren = Entities[currentEntityType].children;
     return zipObject(
       entityChildren,
       entityChildren.map(entityType =>
-        pickBy(entities[entityType], ({ parent }) => parent === entityId),
+        pickBy(
+          entities[entityType],
+          ({ parent, isHidden }) =>
+            parent === entityId && (!isShared || !isHidden),
+        ),
       ),
     );
   },
+);
+
+export const getEmptyHexKeys = createSelector(
+  [getCurrentSector, sidebarEditChildrenSelector],
+  ({ rows, columns }, children = {}) =>
+    difference(
+      allSectorKeys(columns, rows),
+      values(Object.assign({}, ...values(children))).map(({ x, y }) =>
+        coordinateKey(x, y),
+      ),
+    ),
 );
 
 export const isCurrentOrAncestorHidden = createSelector(
