+++ src/store/selectors/sector.selectors.js
@@ -1,15 +1,12 @@
-import { omitBy, difference, values, includes } from 'lodash';
+import { omitBy, includes } from 'lodash';
 import { createSelector } from 'reselect';
 
-import { allSectorKeys, coordinateKey } from 'utils/common';
 import {
   currentSectorSelector,
   sectorSelector,
-  sidebarEditChildrenSelector,
   savedSectorSelector,
   sharedSectorSelector,
 } from 'store/selectors/base.selectors';
-import { getCurrentSector } from 'store/selectors/entity.selectors';
 
 export const getUserSectors = createSelector(
   [sectorSelector, savedSectorSelector],
@@ -28,15 +25,4 @@ export const isCurrentSectorSaved = createSelector(
 export const isViewingSharedSector = createSelector(
   [currentSectorSelector, sharedSectorSelector],
   (currentSector, shared) => includes(shared, currentSector),
-);
-
-export const getEmptyHexKeys = createSelector(
-  [getCurrentSector, sidebarEditChildrenSelector],
-  ({ rows, columns }, children = {}) =>
-    difference(
-      allSectorKeys(columns, rows),
-      values(Object.assign({}, ...values(children))).map(({ x, y }) =>
-        coordinateKey(x, y),
-      ),
-    ),
 );
